Move all code files from top level directory Refactor as needed Rename Dataset.py to lower case
